Move converter into new addon
