+++ sources/webapp/addons/textandbytes/converter/.phpunit.result.cache
@@ -1,0 +1,1 @@
+{"version":1,"defects":{"Tests\\ConverterTest::it_converts_footnotes":4,"Tests\\ConverterTest::it_converts_paragraph_numbers":4,"Tests\\ConverterTest::it_removes_footer":4,"Tests\\ConverterTest::it_converts_sample":4},"times":{"Tests\\ConverterTest::it_converts_footnotes":0.01,"Tests\\ConverterTest::it_converts_paragraph_numbers":0,"Tests\\ConverterTest::it_removes_footer":0.001,"Tests\\ConverterTest::it_converts_sample":0.011}}
