+++ sources/webapp/addons/textandbytes/converter/composer.json
@@ -15,5 +15,16 @@
                 "Textandbytes\\Converter\\ServiceProvider"
             ]
         }
+    },
+    "require-dev": {
+        "statamic/cms": "3.3.51",
+        "phpunit/phpunit": "^9.5",
+        "nunomaduro/collision": "^6.2",
+        "orchestra/testbench": "^7.5"
+    },
+    "config": {
+        "allow-plugins": {
+            "pixelfear/composer-dist-plugin": true
+        }
     }
 }
