+++ sources/webapp/addons/textandbytes/converter/src/ServiceProvider.php
@@ -9,6 +9,14 @@ class ServiceProvider extends AddonServiceProvider
     protected $fieldtypes = [
         Fieldtypes\Converter::class,
     ];
+
+    protected $commands = [
+        Commands\ImportDocuments::class,
+    ];
+    
+    protected $routes = [
+        'cp'  => __DIR__.'/../routes/cp.php',
+    ];
     
     protected $scripts = [
         __DIR__ . '/../dist/js/addon.js',
