+++ sources/webapp/addons/textandbytes/converter/tests/TestCase.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Tests;
+
+use Textandbytes\Converter\ServiceProvider;
+
+class TestCase extends \Orchestra\Testbench\TestCase
+{
+    public function setUp(): void
+    {
+        parent::setUp();
+    }
+
+    protected function getPackageProviders($app)
+    {
+        return [
+            ServiceProvider::class,
+        ];
+    }
+}
