+++ sources/webapp/app/Providers/AppServiceProvider.php
@@ -2,7 +2,6 @@
 
 namespace App\Providers;
 
-use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider;
 use Statamic\Statamic;
 
@@ -27,11 +26,5 @@ class AppServiceProvider extends ServiceProvider
     {
         Statamic::script('app', 'cp.js');
         date_default_timezone_set('Europe/Zurich');
-
-        Statamic::pushCpRoutes(function () {
-            Route::namespace('\App\Http\Controllers')->group(function () {
-                require base_path('routes/cp.php');
-            });
-        });
     }
 }
